requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.2.2
python-dateutil==2.8.2
pydantic==1.10.12
